exclude function names from data breakpoint eligibility
